requirements.txt
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
ipykernel
OpenAI
langchain
ipykernel
langchain_openai
langchain_experimental
ensure
pypdf
reportlab
tweepy
langgraph
elevenlabs